revert: move session token login provider to public package
Refs: 051c5d79ef
