--- cmd/internal/loginprovider/export_test.go
@@ -1,10 +1,0 @@
-// Copyright 2012, 2013, 2022 Canonical Ltd.
-// Licensed under the AGPLv3, see LICENCE file for details.
-
-package loginprovider
-
-var (
-	LoginDeviceAPICall           = &loginDeviceAPICall
-	GetDeviceSessionTokenAPICall = &getDeviceSessionTokenAPICall
-	LoginWithSessionTokenAPICall = &loginWithSessionTokenAPICall
-)
