+++ cmd/internal/loginprovider/tryinorderloginprovider.go
@@ -1,6 +1,7 @@
 // Copyright 2024 Canonical Ltd.
 // Licensed under the AGPLv3, see LICENCE file for details.
 
+// Contains login providers required only by the Juju CLI.
 package loginprovider
 
 import (
