+++ cmd/modelcmd/base.go
@@ -25,7 +25,6 @@ import (
 	"github.com/juju/juju/api/client/modelmanager"
 	k8sproxy "github.com/juju/juju/caas/kubernetes/provider/proxy"
 	"github.com/juju/juju/cloud"
-	"github.com/juju/juju/cmd/internal/loginprovider"
 	"github.com/juju/juju/core/network"
 	"github.com/juju/juju/environs"
 	environscloudspec "github.com/juju/juju/environs/cloudspec"
@@ -593,7 +592,7 @@ func newAPIConnectionParams(
 	dialOpts.BakeryClient = bakery
 
 	if accountDetails.Type == jujuclient.OAuth2DeviceFlowAccountDetailsType {
-		dialOpts.LoginProvider = loginprovider.NewSessionTokenLoginProvider(
+		dialOpts.LoginProvider = api.NewSessionTokenLoginProvider(
 			accountDetails.SessionToken,
 			cmdOut,
 			func(sessionToken string) error {
